# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: EscaShop Queue Management API
  version: "3.1.0"
  description: |
    EscaShop Queue Management System API for managing customer queues, status transitions, and real-time updates.
    
    **Version 3.1.0 Changes (Blue/Green Deployment Ready):**
    - Enhanced 'processing' status support with backward compatibility
    - Blue/Green deployment strategy with zero-downtime updates
    - Forward-compatible database migrations
    - Non-breaking enum value additions
    - Rollback-safe API design
    
    **Deployment Strategy:**
    This API is designed for Blue/Green deployments with:
    - Database migrations applied before backend deployment
    - Backward-compatible enum values
    - Feature flags for UI rollbacks
    - Health check endpoints for monitoring
    
    **Forward Compatibility Note:**
    Clients should ignore unknown future statuses to maintain compatibility with future API versions.
  contact:
    name: EscaShop Development Team
  license:
    name: MIT
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://api.escashop.com/api
    description: Production server
paths:
...
  /customers/{id}/export/excel:
    get:
      summary: Export single customer to Excel
      description: |
        Export a single customer's data to Excel format (.xlsx).
        
        **Field Order Update (v3.1):** Payment Amount and OR Number are now positioned side by side (columns 19-20) for better data visualization and analysis.
        
        **Export Structure:**
        - Columns 1-18: Basic customer information
        - Column 19: **Payment Amount** (Currency format with ₱ symbol)
        - Column 20: **OR Number** (Text format) 
        - Columns 21+: Additional customer details
      tags:
        - Data Export
      parameters:
        - name: id
          in: path
          description: Customer ID to export
          required: true
          schema:
            type: integer
          example: 123
      responses:
        '200':
          description: Excel file generated successfully
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
          headers:
            Content-Disposition:
              description: Attachment filename
              schema:
                type: string
                example: "attachment; filename=customer_123_export.xlsx"
        '404':
          description: Customer not found
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - bearerAuth: []
  /customers/{id}/export/pdf:
    get:
      summary: Export single customer to PDF
      description: |
        Export a single customer's data to PDF format.
        
        **Field Order Update (v3.1):** Payment Amount and OR Number are prominently displayed with enhanced formatting for better visibility.
      tags:
        - Data Export
      parameters:
        - name: id
          in: path
          description: Customer ID to export
          required: true
          schema:
            type: integer
          example: 123
      responses:
        '200':
          description: PDF file generated successfully
          content:
            application/pdf:
              schema:
                type: string
                format: binary
          headers:
            Content-Disposition:
              description: Attachment filename
              schema:
                type: string
                example: "attachment; filename=customer_123_export.pdf"
        '404':
          description: Customer not found
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - bearerAuth: []
  /customers/{id}/export/sheets:
    post:
      summary: Export single customer to Google Sheets
      description: |
        Export a single customer's data to Google Sheets via API integration.
        
        **Field Order Update (v3.1):** Payment Amount and OR Number are configured side by side with proper formatting - Payment Amount as currency, OR Number as text.
      tags:
        - Data Export
      parameters:
        - name: id
          in: path
          description: Customer ID to export
          required: true
          schema:
            type: integer
          example: 123
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                spreadsheetId:
                  type: string
                  description: Google Sheets spreadsheet ID
                  example: "1BxiMVs0XRA5nFMdKvBdBZjgmUUqptlbs74OgvE2upms"
                sheetName:
                  type: string
                  description: Sheet name within the spreadsheet
                  example: "Customer Data"
              required:
                - spreadsheetId
      responses:
        '200':
          description: Data exported to Google Sheets successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  spreadsheetUrl:
                    type: string
                    format: uri
        '404':
          description: Customer not found
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - bearerAuth: []
  /export/excel:
    post:
      summary: Export multiple customers to Excel
      description: |
        Export multiple customers' data to Excel format (.xlsx).
        
        **Field Order Update (v3.1):** Payment Amount and OR Number are positioned side by side (columns 19-20) across all customer records for consistent data analysis.
      tags:
        - Data Export
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                customerIds:
                  type: array
                  items:
                    type: integer
                  description: Array of customer IDs to export
                  example: [123, 124, 125]
                filters:
                  type: object
                  description: Optional filters for customer selection
                  properties:
                    status:
                      $ref: '#/components/schemas/QueueStatus'
                    dateRange:
                      type: object
                      properties:
                        start:
                          type: string
                          format: date
                        end:
                          type: string
                          format: date
      responses:
        '200':
          description: Excel file with multiple customers generated successfully
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
          headers:
            Content-Disposition:
              description: Attachment filename
              schema:
                type: string
                example: "attachment; filename=customers_export.xlsx"
        '400':
          description: Invalid request body or no customers found
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - bearerAuth: []
  /export/pdf:
    post:
      summary: Export multiple customers to PDF
      description: |
        Export multiple customers' data to PDF format in table layout.
        
        **Field Order Update (v3.1):** Payment Amount and OR Number are positioned adjacently in the table for easy comparison and analysis.
      tags:
        - Data Export
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                customerIds:
                  type: array
                  items:
                    type: integer
                  description: Array of customer IDs to export
                  example: [123, 124, 125]
                filters:
                  type: object
                  description: Optional filters for customer selection
                  properties:
                    status:
                      $ref: '#/components/schemas/QueueStatus'
                    dateRange:
                      type: object
                      properties:
                        start:
                          type: string
                          format: date
                        end:
                          type: string
                          format: date
      responses:
        '200':
          description: PDF file with multiple customers generated successfully
          content:
            application/pdf:
              schema:
                type: string
                format: binary
          headers:
            Content-Disposition:
              description: Attachment filename
              schema:
                type: string
                example: "attachment; filename=customers_export.pdf"
        '400':
          description: Invalid request body or no customers found
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - bearerAuth: []
...
components:
  schemas:
    QueueStatus:
      type: string
      enum:
        - waiting
        - serving
        - processing
        - completed
        - cancelled
      description: |
        Queue status enumeration.
        
        **v1.1 Addition:** Added 'processing' status for enhanced workflow support.
        
        **Status Descriptions:**
        - `waiting`: Customer is in queue waiting to be called
        - `serving`: Customer is currently being served at a counter
        - `processing`: Customer's order is being prepared/processed (v1.1)
        - `completed`: Customer service has been completed
        - `cancelled`: Customer service was cancelled
        
        **Forward Compatibility Note:** Clients should gracefully handle and ignore
        any unknown status values that may be added in future API versions.
      example: processing
...
  responses:
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Authentication required"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Internal server error"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...